little fixes
- important fix in list.h
+ other little fixes

+++ src/words.h
@@ -44,7 +44,7 @@ int zwischen(const char text[], const char word1[], const char word2[], char sav
 
 int intLen(int zahl);
 void intToString(int zahl, char * saveto);
-void doubleToString(float zahl, int afterpoint, char * saveto);
+void doubleToString(double zahl, int afterpoint, char * saveto);
 
 const char * word_seek(const char * word, const char * search);
 const char * word_seek_after(const char * word, const char * search);
@@ -563,7 +563,7 @@ void intToString(int zahl, char * saveto) {
     if (intLen(next) == 0) { saveto++; *saveto = '\0'; return; }
     intToString(next, ++saveto);
 }
-void doubleToString(float zahl, int afterpoint, char * saveto) {
+void doubleToString(double zahl, int afterpoint, char * saveto) {
     int ipart = (int) zahl; // Extract integer part
     double fpart = zahl - (double) ipart; // Extract floating part == part after the comma
